requirements.txt
pyyaml
PyQt5
praw
pillow